+++ builder/src/test/java/com/stratio/cassandra/lucene/builder/BuilderTest.java
@@ -613,15 +613,18 @@ public class BuilderTest {
 
     @Test
     public void testGeoShapeConditionFull() {
-        String actual = geoShape("field", "shape").operation("intersects")
-                                                  .transform(bufferTransformation().maxDistance("2km"),
-                                                             bufferTransformation().minDistance("1km"),
-                                                             bufferTransformation().maxDistance("10km")
-                                                                                   .minDistance("5km"))
-                                                  .build();
-        String expected = "{\"type\":\"geo_shape\",\"field\":\"field\",\"shape\":\"shape\"," +
-                          "\"operation\":\"intersects\",\"transformations\":[{\"type\":\"buffer\"," +
-                          "\"max_distance\":\"2km\"},{\"type\":\"buffer\",\"min_distance\":\"1km\"}," +
+        String actual = geoShape("field", "my_shape").operation("intersects")
+                                                     .transform(differenceGeoTransformation("my_difference_shape"),
+                                                                intersectionGeoTransformation("my_intersection_shape"),
+                                                                unionGeoTransformation("my_union_shape"),
+                                                                bufferGeoTransformation().maxDistance("10km")
+                                                                                         .minDistance("5km"))
+                                                     .build();
+        String expected = "{\"type\":\"geo_shape\",\"field\":\"field\",\"shape\":\"my_shape\"," +
+                          "\"operation\":\"intersects\",\"transformations\":[" +
+                          "{\"type\":\"difference\",\"shape\":\"my_difference_shape\"}," +
+                          "{\"type\":\"intersection\",\"shape\":\"my_intersection_shape\"}," +
+                          "{\"type\":\"union\",\"shape\":\"my_union_shape\"}," +
                           "{\"type\":\"buffer\",\"max_distance\":\"10km\",\"min_distance\":\"5km\"}]}";
         assertEquals("geo shape condition serialization is wrong", expected, actual);
     }