+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * Add search by WKT geographical shapes (point, linestring, polygon and their multipart)
 * Add API for search-time transformation of WKT geographical shapes
 * Add transformation for getting the buffer around a geographical shape
+* Add transformation for getting the difference between two geographical shapes
+* Add transformation for getting the intersection between two geographical shapes
+* Add transformation for getting the union between two geographical shapes
 * Fix geo distance parsing of nautical miles
 
 ## 2.2.4.3 (19 February 2016)